reorganize component and callbacks modules
